backend/core-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
redis>=5.0.0
pyotp>=2.9.0
qrcode>=7.4.2
email-validator>=2.0.0
twilio>=9.0.0
sendgrid>=6.10.0
